Factored out individual test suites into their own files
