+++ test/mocha.opts
@@ -1,4 +1,3 @@
---recursive
 --require babel-core/register
 --exit
 --bail
